fix(amazonq): Fix Generate unit tests from right click context menu (#5705)
* initial changes

* fix gen unit tests
